pyproject.toml
[project]
name = "music-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "google-adk>=1.25.1",
  "fastapi[standard]>=0.129.2",
  "uvicorn>=0.41.0",
  "httpx>=0.28.1",
  "python-dotenv>=1.2.1",
  "litellm>=1.81.14",
  "ruff>=0.15.2",
]
